requirements.txt
alembic==1.9.1
bcrypt==4.0.1
blinker==1.5
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
elastic-transport==8.4.0
elasticsearch==8.5.3
email-validator==1.3.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
SQLAlchemy==1.4.45
urllib3==1.26.13
Werkzeug==2.2.2
WTForms==3.0.1